CS2-DatasetCollector/requirements.txt
acme==2.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
attrs==23.2.0
Babel==2.10.3
bcc==0.29.1
blinker==1.7.0
Brlapi==0.8.5
certbot==2.9.0
certifi==2023.11.17
chardet==5.2.0
click==8.1.6
cloud-init==25.2
colorama==0.4.6
command-not-found==0.3
ConfigArgParse==1.7
configobj==5.0.8
cryptography==41.0.7
cupshelpers==1.0
dbus-python==1.3.2
defer==1.0.6
distro==1.9.0
distro-info==1.7+build1
evdev==1.9.2
frozenlist==1.8.0
httplib2==0.20.4
idna==3.6
Jinja2==3.1.2
josepy==1.14.0
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
louis==3.29.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
menulibre==2.4.0
mss==10.1.0
multidict==6.7.0
netaddr==0.8.0
numpy==2.2.6
oauthlib==3.2.2
olefile==0.46
opencv-python==4.12.0.88
parsedatetime==2.6
pexpect==4.9.0
pillow==10.2.0
propcache==0.4.1
psutil==5.9.8
ptyprocess==0.7.0
PyAudio==0.2.13
pycairo==1.25.1
pycups==2.0.1
Pygments==2.17.2
PyGObject==3.48.2
PyICU==2.12
PyJWT==2.7.0
pynput==1.8.1
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyRFC3339==1.1
pyrsistent==0.20.0
pyserial==3.5
python-apt==2.7.7+ubuntu5
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
python-xlib==0.33
pytz==2024.1
pyxdg==0.28
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
screen-resolution-extra==0.0.0
setuptools==68.1.2
six==1.16.0
systemd-python==235
typing_extensions==4.10.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.2
unattended-upgrades==0.1
urllib3==2.0.7
wadllib==1.3.6
wheel==0.42.0
xdg==5
xkit==0.0.0
yarl==1.22.0
